requirements.txt
PySide6==6.11.0
PySide6-Addons==6.11.0
Pillow==12.2.0
